chore(ci): remove codeql action (#1027)
This old action has the following error in the logs:

> Error: This version of the CodeQL Action was deprecated on January
18th, 2023, and is no longer updated or supported. For better
performance, improved security, and new features, upgrade to v2. For
more information, see
https://github.blog/changelog/2023-01-18-code-scanning-codeql-action-v1-is-now-deprecated/

I'm going to remove and see if there is an easier way to add v2.

Or maybe its no longer necessary since socket.dev is basically the same
thing.